# api/openapi.yaml
openapi: 3.0.3
info:
  title: rffmpeg API
  description: Remote ffmpeg transcoding service API
  version: 1.0.0
  contact:
    name: rffmpeg
    url: https://github.com/tsix404/rffmpeg
servers:
  - url: /api/v1
    description: Default server
...
paths:
...
  /upload/init:
    post:
      tags: [files]
      summary: Initialize chunked upload
      description: Initialize a new chunked upload session for large files
      operationId: initChunkUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChunkUploadInitRequest'
      responses:
        '200':
          description: Upload session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunkUploadInitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /upload/chunk/{uploadId}/{chunkIndex}:
    post:
      tags: [files]
      summary: Upload a chunk
      description: Upload a single chunk of a chunked upload session
      operationId: uploadChunk
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
        - name: chunkIndex
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [chunk]
              properties:
                chunk:
                  type: string
                  format: binary
                checksum:
                  type: string
                  description: Optional SHA256 checksum of the chunk
      responses:
        '200':
          description: Chunk uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunkUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /upload/progress/{uploadId}:
    get:
      tags: [files]
      summary: Get upload progress
      description: Get the progress of a chunked upload session
      operationId: getUploadProgress
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Upload progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadProgressResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /upload/complete:
    post:
      tags: [files]
      summary: Complete chunked upload
      description: Finalize a chunked upload and assemble the file
      operationId: completeChunkUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChunkUploadCompleteRequest'
      responses:
        '200':
          description: File assembled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunkUploadCompleteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /upload/cancel/{uploadId}:
    post:
      tags: [files]
      summary: Cancel chunked upload
      description: Cancel a chunked upload session and clean up temporary files
      operationId: cancelChunkUpload
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Upload cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /upload/resume/{uploadId}:
    get:
      tags: [files]
      summary: Resume chunked upload
      description: Get information needed to resume an interrupted chunked upload
      operationId: resumeChunkUpload
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Upload resume information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResumeResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    ChunkUploadInitRequest:
      type: object
      required: [filename, file_size]
      properties:
        filename:
          type: string
        file_size:
          type: integer
          format: int64
        chunk_size:
          type: integer
          format: int64
          description: Optional chunk size in bytes (default 10MB)
        total_chunks:
          type: integer
          description: Optional total chunk count (calculated from file_size/chunk_size)
        checksum:
          type: string
          description: Optional SHA256 checksum of entire file for final verification
    ChunkUploadInitResponse:
      type: object
      required: [upload_id, chunk_size, total_chunks]
      properties:
        upload_id:
          type: string
        chunk_size:
          type: integer
          format: int64
        total_chunks:
          type: integer
        message:
          type: string
    ChunkUploadResponse:
      type: object
      required: [upload_id, chunk_index]
      properties:
        upload_id:
          type: string
        chunk_index:
          type: integer
        message:
          type: string
    UploadProgressResponse:
      type: object
      required: [upload_id, filename, file_size, chunk_size, total_chunks, status]
      properties:
        upload_id:
          type: string
        filename:
          type: string
        file_size:
          type: integer
          format: int64
        chunk_size:
          type: integer
          format: int64
        total_chunks:
          type: integer
        uploaded_chunks:
          type: integer
        uploaded_size:
          type: integer
          format: int64
        completed_chunks:
          type: array
          items:
            type: integer
        status:
          type: string
          enum: [in_progress, completed, expired, cancelled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChunkUploadCompleteRequest:
      type: object
      required: [upload_id]
      properties:
        upload_id:
          type: string
        checksum:
          type: string
          description: Optional SHA256 checksum for final verification
    ChunkUploadCompleteResponse:
      type: object
      required: [file_id]
      properties:
        file_id:
          type: string
        message:
          type: string
    UploadResumeResponse:
      type: object
      properties:
        upload_id:
          type: string
        filename:
          type: string
        chunk_size:
          type: integer
          format: int64
        total_chunks:
          type: integer
        completed_chunks:
          type: array
          items:
            type: integer
        missing_chunks:
          type: array
          items:
            type: integer
        status:
          type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...